requirements.txt
asgiref==3.9.0
Django==5.2.4
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.3
